fix: make Integration suite green on PHP 8.4 with real MySQL
PHP 8.4 compatibility across core models, controllers, utils, and exceptions:
type casts for strict internal functions, public method visibility for test
access, and BC alias registration after oxNew for is_subclass_of checks.

Fix StrMb::preg_replace array replacement (SEO URL corruption), image
generator int casts in oxpicgenerator, Symfony mailer body fallback, and
admin Language tpl fallback when theme lang files are absent.

Add oe:database:reset console command, admin_twig menu fixture, PHP built-in
server router, fake sendmail for mail tests, and integration test stubs.

+++ source/Application/Model/Voucher.php
@@ -340,8 +340,8 @@ class Voucher extends \OxidEsales\Eshop\Core\Model\BaseModel
         $iTime = time();
 
         // If date is not set will add day before and day after to check if voucher valid today.
-        $iTomorrow = mktime(0, 0, 0, date('m'), date('d') + 1, date('Y'));
-        $iYesterday = mktime(0, 0, 0, date('m'), date('d') - 1, date('Y'));
+        $iTomorrow = mktime(0, 0, 0, (int) date('m'), (int) date('d') + 1, (int) date('Y'));
+        $iYesterday = mktime(0, 0, 0, (int) date('m'), (int) date('d') - 1, (int) date('Y'));
 
         // Checks if beginning date is set, if not set $iFrom to yesterday so it will be valid.
         $iFrom = ((int) $oSeries->oxvoucherseries__oxbegindate->value) ?
